Refactor export happiness to use jobs
