+++ happiness-backend/api/templates/happiness_export.html
@@ -1,0 +1,7 @@
+<p>Dear {{ user.username }},</p>
+<p>
+    Your happiness export is attached to this email.
+</p>
+<p>Enjoy viewing your data!</p>
+<p>Sincerely,</p>
+<p>Happiness App</p>
